authorization added and view cleanups

+++ routes/web.php
@@ -1,8 +1,5 @@
 <?php
-
-use App\Models\Comment;
 use Illuminate\Support\Facades\Route;
-
 use App\Http\Controllers\LikeController;
 use App\Http\Controllers\TweetController;
 use App\Http\Controllers\CommentController;
@@ -15,38 +12,46 @@ use App\Http\Controllers\TrendingTweetsController;
 use App\Http\Controllers\UserProfileController;
 
 Route::redirect('/', '/tweets');
+Route::view('/about', 'about');
 
+//Resource Controllers
+// Public routes
 Route::resource('tweets', TweetController::class);
 
-Route::get('/about', function() {
-    return view('about');
-});
+Route::resource('users', UserProfileController::class)->only('show');
+Route::resource('users', UserProfileController::class)
+    ->only(['edit', 'update'])
+    ->middleware('auth');
 
-// Route::get('/users/{user}', [ViewUserController::class, 'show']);
-// Route::get('/users/{user}/edit', [ViewUserController::class, 'edit']);
+Route::resource('tweets.comments', CommentController::class)
+    ->only(['create', 'store'])
+    ->middleware('auth');
+Route::resource('tweets.comments', CommentController::class)
+    ->only('destroy')
+    ->middleware('auth');
 
-Route::resource('users', UserProfileController::class)->only(['show', 'edit', 'update']);
-
-
+//Trending Tweets
 Route::get('/trending', [TrendingTweetsController::class, 'index']);
 
-Route::get('/users/{user}/feed', [FeedController::class, 'index']);
+//User feeds
+Route::get('/users/{user}/feed', [FeedController::class, 'index'])->middleware('auth');
 
-Route::resource('tweets.comments', CommentController::class)->only(['create', 'store', 'destroy']);
+//Auth 
+Route::middleware('guest')->group( function() {
+    Route::get('/register', [RegisteredUserController::class, 'create'])->name('register');
+    Route::post('/register', [RegisteredUserController::class, 'store']);
 
+    Route::get('/login', [SessionController::class, 'create'])->name('login');
+    Route::post('/login', [SessionController::class, 'store']);
 
-//Auth
-Route::get('/register', [RegisteredUserController::class, 'create']);
-Route::post('/register', [RegisteredUserController::class, 'store']);
+});
 
-Route::get('/login', [SessionController::class, 'create']);
-Route::post('/login', [SessionController::class, 'store']);
-Route::delete('/logout', [SessionController::class, 'destroy']);
+Route::delete('/logout', [SessionController::class, 'destroy'])->middleware('auth');
 
 //Like and Dislike
-Route::post('/tweets/{tweet}/like', [LikeController::class, 'store']);
-Route::post('/tweets/{tweet}/dislike', [DislikeController::class, 'store']);
+Route::post('/tweets/{tweet}/like', [LikeController::class, 'store'])->name('tweets.like')->middleware('auth');
+Route::post('/tweets/{tweet}/dislike', [DislikeController::class, 'store'])->name('tweets.dislike')->middleware('auth');
 
 //Follow and Unfollow
-Route::post('/users/{user}/follow', [FollowController::class, 'store']);
-Route::delete('/users/{user}/unfollow', [FollowController::class, 'destroy']);
+Route::post('/users/{user}/follow', [FollowController::class, 'store'])->name('user.follow')->middleware('auth');
+Route::delete('/users/{user}/unfollow', [FollowController::class, 'destroy'])->name('user.unfollow')->middleware('auth');
